+++ testutil/keeper/unit_test_helpers.go
@@ -232,12 +232,15 @@ func SetupForStoppingConsumerChain(t *testing.T, ctx sdk.Context,
 	gomock.InOrder(expectations...)
 
 	providerKeeper.SetConsumerChainId(ctx, consumerId, "chainID")
-	providerKeeper.SetConsumerMetadata(ctx, consumerId, GetTestConsumerMetadata())
-	providerKeeper.SetConsumerInitializationParameters(ctx, consumerId, GetTestInitializationParameters())
-	providerKeeper.SetConsumerPowerShapingParameters(ctx, consumerId, GetTestPowerShapingParameters())
+	err := providerKeeper.SetConsumerMetadata(ctx, consumerId, GetTestConsumerMetadata())
+	require.NoError(t, err)
+	err = providerKeeper.SetConsumerInitializationParameters(ctx, consumerId, GetTestInitializationParameters())
+	require.NoError(t, err)
+	err = providerKeeper.SetConsumerPowerShapingParameters(ctx, consumerId, GetTestPowerShapingParameters())
+	require.NoError(t, err)
 	providerKeeper.SetConsumerPhase(ctx, consumerId, providertypes.ConsumerPhase_CONSUMER_PHASE_INITIALIZED)
 
-	err := providerKeeper.CreateConsumerClient(ctx, consumerId)
+	err = providerKeeper.CreateConsumerClient(ctx, consumerId)
 	require.NoError(t, err)
 	// set the mapping consumer ID <> client ID for the consumer chain
 	providerKeeper.SetConsumerClientId(ctx, consumerId, "clientID")
