+++ x/ccv/types/errors.go
@@ -22,4 +22,6 @@ var (
 	ErrDuplicateConsumerChain      = errorsmod.Register(ModuleName, 14, "consumer chain already exists")
 	ErrConsumerChainNotFound       = errorsmod.Register(ModuleName, 15, "consumer chain not found")
 	ErrInvalidDoubleVotingEvidence = errorsmod.Register(ModuleName, 16, "invalid consumer double voting evidence")
+	ErrStoreKeyNotFound            = errorsmod.Register(ModuleName, 17, "store key not found")
+	ErrStoreUnmarshal              = errorsmod.Register(ModuleName, 18, "cannot unmarshal value from store")
 )
